+++ ft_memset.c
@@ -6,14 +6,16 @@
 /*   By: tjorge-l <tjorge-l@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/04/12 12:47:27 by tjorge-l          #+#    #+#             */
-/*   Updated: 2024/04/12 12:47:28 by tjorge-l         ###   ########.fr       */
+/*   Updated: 2024/04/17 15:43:21 by tjorge-l         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-void	*ft_memset(void *s, int c, unsigned int n)
+#include "libft.h"
+
+void	*ft_memset(void *s, int c, size_t n)
 {
-	unsigned int		i;
-	char				*str;
+	size_t		i;
+	char		*str;
 
 	str = s;
 	i = 0;
